moved functions that generate chart data into their own file, now imported in GenerateCharts component
