+++ client/src/components/GenerateCharts.jsx
@@ -1,9 +1,8 @@
 import { useEffect, useState } from "react";
 import MonthDropDown from "./YearMonthDropDown";
 import { Chart } from "react-google-charts";
-import { convertToNumber } from "./handleStringNumbers";
 import GenerateTables from "./GenTable";
-import { getMonthNum } from "./handleDates";
+import { generateData, generateLineChartData } from "./genChartDataFunctions";
 
 const GenerateCharts = ({
   expenses,
@@ -36,7 +35,7 @@ const GenerateCharts = ({
 
   const columnChartOptions = {
     title: `Monthly Spendings and Income (${year})`,
-    vAxis: { title: "in Dollars (USD)" },
+    vAxis: { title: "Amount" },
     hAxis: { title: "Month" },
     seriesType: "bars",
     series: { 2: { type: "line" } },
@@ -100,76 +99,3 @@ const GenerateCharts = ({
 };
 
 export default GenerateCharts;
-
-/**
- * Generates the data needed for the Charts by iterating through the expense array
- * and summing the total of each expense tag.
- * @param {array} expenseArray - Array of expense objects for the month + year selected
- * @returns an array of sub arrays
- */
-function generateData(expenseArray) {
-  // object to hold the total of users spending by category
-  const dataObj = {};
-  expenseArray.forEach((expenseObj) => {
-    // store the current cateogry
-    let category = expenseObj["tags"];
-    // checks to see if the category has already been saved to either add or
-    // create a new key-value in data obj
-    if (dataObj.hasOwnProperty(category) === true) {
-      dataObj[category] += convertToNumber(expenseObj.amount);
-    } else {
-      dataObj[category] = convertToNumber(expenseObj.amount);
-    }
-  });
-  // returns as an array of sub-arrays made from data obj
-  const expenseData = Object.entries(dataObj);
-  // append the chart data to the beginning of array
-  expenseData.unshift(["Expenses", "Dollars"]);
-  return expenseData;
-}
-
-function generateYearlyData(data) {
-  const result = Array(12).fill(0);
-  if (data.length === 0) return result;
-
-  data.forEach((obj) => {
-    const index = Number(getMonthNum(obj.month)) - 1;
-    result[index] += convertToNumber(obj.amount);
-  });
-
-  return result;
-}
-
-function generateLineChartData(expensesData, incomeData) {
-  let months = [
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
-  ];
-
-  let expenses = generateYearlyData(expensesData);
-  let incomes = generateYearlyData(incomeData);
-
-  let dataRotated = [];
-  months.forEach((currMonth, index) =>
-    dataRotated.push([
-      currMonth,
-      incomes[index],
-      expenses[index],
-      incomes[index] - expenses[index],
-    ])
-  );
-
-  // add average
-  dataRotated.unshift(["Month", "Income", "Expenses", "Money Left"]);
-  return dataRotated;
-}
